requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
distro==1.9.0
exa-py==2.4.0
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.13
langchain-openai==1.1.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
multitasking==0.0.12
numpy==2.4.2
openai==2.21.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.0
peewee==3.19.0
platformdirs==4.9.2
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
websockets==16.0
xxhash==3.6.0
yfinance==1.2.0
zstandard==0.25.0
